rename file to match existing pattern
