+++ src/lib/providers/fileCompletion.ts
